requirements.txt
torch
torchvision
tqdm
numpy
scikit-learn
Pillow
transformers
umap-learn
matplotlib
pandas
seaborn 